+++ common/src/main/kotlin/com/cobblepalsworld/behavior/behaviors/VoidBehavior.kt
@@ -8,7 +8,7 @@ import com.cobblepalsworld.config.ConfigManager
 import com.cobblepalsworld.inventory.InventoryManager
 import com.cobblepalsworld.inventory.PokemonInventory
 import com.cobblepalsworld.navigation.ContainerFinder
-import com.cobblepalsworld.pasture.PastureWorkerManager
+import com.cobblepalsworld.persistence.CobblePalsSaveData
 import com.cobblepalsworld.tag.TagInstance
 import com.cobblepalsworld.tag.TagType
 import com.cobblepalsworld.tag.filter.FilterMatcher
@@ -82,7 +82,7 @@ object VoidBehavior : TagBehavior {
 
         container.markDirty()
         if (extracted > 0) {
-            PastureWorkerManager.markDirtyNow(world)
+            (world as? net.minecraft.server.world.ServerWorld)?.let(CobblePalsSaveData::markDirty)
         }
         return WorkResult.Done()
     }
@@ -97,7 +97,7 @@ object VoidBehavior : TagBehavior {
         }
 
         if (changed) {
-            PastureWorkerManager.markDirtyNow(world)
+            (world as? net.minecraft.server.world.ServerWorld)?.let(CobblePalsSaveData::markDirty)
         }
         return WorkResult.Done()
     }